+++ applications/FluidDynamicsApplication/custom_python/add_custom_strategies_to_python.cpp
@@ -33,7 +33,7 @@
 #include "custom_strategies/strategies/fs_strategy.h"
 
 //schemes
-#include "custom_strategies/schemes/gear_scheme.h"
+#include "custom_strategies/schemes/bdf2_turbulent_scheme.h"
 #include "custom_strategies/schemes/residualbased_simple_steady_scheme.h"
 #include "custom_strategies/schemes/residualbased_predictorcorrector_velocity_bossak_scheme_turbulent.h"
